deletion confirm popup

+++ admin/includes/view_all_users.php
@@ -1,7 +1,7 @@
 
 <div class="card">
   <div class="card-header">
-    <h3>View All Comments</h3>
+    <h2>View All Users</h2>
   </div>
   <div class="card-body">
 
@@ -68,7 +68,7 @@
             // }
 
             echo "<td><a href='users.php?source=edit_user&u_id=$user_id'>Edit</a></td>";
-            echo "<td><a href='users.php?delete=$user_id'>Delete</a></td>";
+            echo "<td><a onClick=\" javascript: return confirm ('Are you sure you want to delete?'); \" href='users.php?delete=$user_id'>Delete</a></td>";
             echo "</tr>";
         }
